Fixed IMU sensor measurements.
Recommended approach is to set Vehicle tag to the top rigidbody of the vessel.
Scripts search for Vehicle tag to be used as rigidbody.

If IMU is added as separate rigidbody on a boat, fixed joint should connect that rigidbody to boat rigidbody.
IMU rigidbody mass should be set to low value, drags should be zerp, and gravity and kinematic flags shoul not be set.
If IMU script is added directly to boat rigidbody, IMU will get measurements from that rigidbody.

If Vehicle tag is not found, treat top rigidbody in hierarchy as vehicle rigidbody.

Added utility function to get top component in hierarchy of parent gameObjects.